requirements.txt
# ML Malware Generator - Dependencies

# Deep Learning Frameworks
torch>=2.0.0
tensorflow>=2.13.0
transformers>=4.30.0

# Reinforcement Learning
gym>=0.26.0
stable-baselines3>=2.0.0

# Assembly & Binary Analysis
capstone>=5.0.0
keystone-engine>=0.9.2
pefile>=2023.2.7
lief>=0.13.0

# Detection & Analysis
yara-python>=4.3.1
pycryptodome>=3.18.0
python-magic>=0.4.27

# API & Web
requests>=2.31.0
aiohttp>=3.8.5

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging & Monitoring
tensorboard>=2.13.0
wandb>=0.15.0
loguru>=0.7.0

# CLI & Utils
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
